+++ new-era-commands/rotations/triage.js
@@ -1,3 +1,3 @@
-const { rotationBuilder } = require("../../services/RotationBuilder");
+const { rotationBuilder } = require("../../utils/rotation-builder");
 
 module.exports = rotationBuilder("triage", "maintainerTriageRotationList");
