+++ services/rotations/index.js
@@ -1,0 +1,1 @@
+export * from "./rotation.service";
